coin end point

+++ src/Data/Entities/Coin.cs
@@ -3,7 +3,7 @@ namespace Data.Entities;
 public class Coin
 {
     public int Id { get; set; }
-    public double InGold { get; set; }
+    public decimal InGold { get; set; }
     public int Count { get; set; }
     public string Name { get; set; }
 }
